+++ example/model/gen.object.user.blogs.go
@@ -393,7 +393,11 @@ func (m *_UserBlogsDBMgr) FetchBySQL(q string, args ...interface{}) (results []*
 }
 
 func (m *_UserBlogsDBMgr) FetchBySQLContext(ctx context.Context, q string, args ...interface{}) (results []*UserBlogs, err error) {
-	rows, err := m.db.QueryContext(ctx, q, args...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return nil, fmt.Errorf("%s", "db has no context")
+	}
+	rows, err := ctxDB.QueryContext(ctx, q, args...)
 	if err != nil {
 		return nil, fmt.Errorf("UserBlogs fetch error: %v", err)
 	}
@@ -661,7 +665,11 @@ func (m *_UserBlogsDBMgr) queryLimit(where string, limit int, args ...interface{
 func (m *_UserBlogsDBMgr) queryLimitContext(ctx context.Context, where string, limit int, args ...interface{}) (results []PrimaryKey, err error) {
 	pk := UserBlogsMgr.NewPrimaryKey()
 	query := fmt.Sprintf("SELECT %s FROM user_blogs %s", strings.Join(pk.Columns(), ","), where)
-	rows, err := m.db.QueryContext(ctx, query, args...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return nil, fmt.Errorf("%s", "db has no context")
+	}
+	rows, err := ctxDB.QueryContext(ctx, query, args...)
 	if err != nil {
 		return nil, fmt.Errorf("UserBlogs query limit error: %v", err)
 	}
@@ -712,7 +720,11 @@ func (m *_UserBlogsDBMgr) queryCount(where string, args ...interface{}) (int64, 
 
 func (m *_UserBlogsDBMgr) queryCountContext(ctx context.Context, where string, args ...interface{}) (int64, error) {
 	query := fmt.Sprintf("SELECT count(`user_id`) FROM user_blogs %s", where)
-	rows, err := m.db.QueryContext(ctx, query, args...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	rows, err := ctxDB.QueryContext(ctx, query, args...)
 	if err != nil {
 		return 0, fmt.Errorf("UserBlogs query count error: %v", err)
 	}
@@ -762,7 +774,11 @@ func (m *_UserBlogsDBMgr) BatchCreateContext(ctx context.Context, objs []*UserBl
 		values = append(values, obj.BlogId)
 	}
 	query := fmt.Sprintf("INSERT INTO user_blogs(%s) VALUES %s", strings.Join(objs[0].GetNoneIncrementColumns(), ","), strings.Join(params, ","))
-	result, err := m.db.ExecContext(ctx, query, values...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	result, err := ctxDB.ExecContext(ctx, query, values...)
 	if err != nil {
 		return 0, err
 	}
@@ -794,7 +810,11 @@ func (m *_UserBlogsDBMgr) UpdateBySQLContext(ctx context.Context, set, where str
 	if where != "" {
 		query = fmt.Sprintf("UPDATE user_blogs SET %s WHERE %s", set, where)
 	}
-	result, err := m.db.ExecContext(ctx, query, args...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	result, err := ctxDB.ExecContext(ctx, query, args...)
 	if err != nil {
 		return 0, err
 	}
@@ -826,7 +846,11 @@ func (m *_UserBlogsDBMgr) CreateContext(ctx context.Context, obj *UserBlogs) (in
 	values := make([]interface{}, 0, 2)
 	values = append(values, obj.UserId)
 	values = append(values, obj.BlogId)
-	result, err := m.db.ExecContext(ctx, q, values...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	result, err := ctxDB.ExecContext(ctx, q, values...)
 	if err != nil {
 		return 0, err
 	}
@@ -856,7 +880,11 @@ func (m *_UserBlogsDBMgr) UpdateContext(ctx context.Context, obj *UserBlogs) (in
 	values := make([]interface{}, 0, 2-2)
 	values = append(values, pk.SQLParams()...)
 
-	result, err := m.db.ExecContext(ctx, q, values...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	result, err := ctxDB.ExecContext(ctx, q, values...)
 	if err != nil {
 		return 0, err
 	}
@@ -911,8 +939,12 @@ func (m *_UserBlogsDBMgr) DeleteByPrimaryKeyContext(ctx context.Context, userId 
 		UserId: userId,
 		BlogId: blogId,
 	}
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
 	q := fmt.Sprintf("DELETE FROM user_blogs %s", pk.SQLFormat())
-	result, err := m.db.ExecContext(ctx, q, pk.SQLParams()...)
+	result, err := ctxDB.ExecContext(ctx, q, pk.SQLParams()...)
 	if err != nil {
 		return 0, err
 	}
@@ -936,7 +968,11 @@ func (m *_UserBlogsDBMgr) DeleteBySQLContext(ctx context.Context, where string, 
 	if where != "" {
 		query = fmt.Sprintf("DELETE FROM user_blogs WHERE %s", where)
 	}
-	result, err := m.db.ExecContext(ctx, query, args...)
+	ctxDB, ok := m.db.(orm.ContextDB)
+	if !ok {
+		return 0, fmt.Errorf("%s", "db has no context")
+	}
+	result, err := ctxDB.ExecContext(ctx, query, args...)
 	if err != nil {
 		return 0, err
 	}
